Add --typedRefPropViews for some workaround regarding circular types
Implements the trick mentioned in this comment:

https://github.com/mobxjs/mobx-state-tree/issues/943#issuecomment-442603532

Every late() reference field will be of type IAnyModelType and will be
accompanied by a view getter casting it to the actual type of the prop.
After this a reference field name "refField" should be accessed as
"refFieldProp" which will provide correct typings/autocomplete in IDEs.

This trick works up to typescript 3.4.x and doesn't work in 3.5 and
above :-( Using Instance<...> like this however still works in all
versions:

https://github.com/mobxjs/mobx-state-tree/issues/943#issuecomment-442528849

refs #45